chore: rename patch to content
