+++ src/components/SystemsView/columns/content/cells/Template.tsx
